+++ src/main/java/edu/progmatic/blood_pressure_diary/repositories/UserRepository.java
@@ -11,7 +11,7 @@ public interface UserRepository extends JpaRepository<User, Integer> {
     User findByUsername(String username);
     User findByEmail(String email);
     @Query("SElECT user FROM  User user WHERE user.id=:id")
-    User findUserById(@Param(value = "id")Integer id);
+    User findUserById(@Param(value = "id")Long id);
 //    List<User> findByUsernameContaining(String username);
